Commit#2
In the NV  on line 6 i've create a object of type Item which is the class i want to test, on line 10 in the NV i create an instance of the of the class.
fom line 17 to 18 i try to test the first method to see if the cost will be the same as what i expect.
same as on line 23 and 24 in do the same to see if the exchange will be as i expect